requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
ephem==3.7.6.0
Flask==1.1.1
geoip2==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Logbook==1.4.3
MarkupSafe==1.1.1
maxminddb==1.4.1
mysql-connector-python==8.0.16
protobuf==3.9.0
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
ratelimiter==1.2.0.post0
Represent==1.6.0
requests==2.22.0
six==1.12.0
spacetrack==0.13.3
SQLAlchemy==1.3.6
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.2
Werkzeug==0.15.4
